# api/openapi.yaml
openapi: 3.0.3
info:
  title: Recovery Plus API
  description: Production-grade backend for athletic recovery automation
  version: 1.0.0
  contact:
    name: Recovery Plus Team
    email: api@recoveryplus.io
servers:
  - url: http://localhost:3000
    description: Local development
  - url: https://api.recoveryplus.io
    description: Production
security:
  - bearerAuth: []
...
paths:
...
  /v1/surveys/session:
    post:
      tags: [Surveys]
      summary: Submit session survey
      description: Records post-session subjective metrics
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SessionSurveyRequest'
      responses:
        '201':
          description: Survey recorded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionSurveyResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /v1/surveys/history:
    get:
      tags: [Surveys]
      summary: Get survey history
      parameters:
        - name: from
          in: query
          schema:
            type: string
            format: date
        - name: to
          in: query
          schema:
            type: string
            format: date
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
      responses:
        '200':
          description: Survey history
          content:
            application/json:
              schema:
                type: object
                properties:
                  surveys:
                    type: array
                    items:
                      $ref: '#/components/schemas/SessionSurvey'
  /v1/prescriptions/today:
    get:
      tags: [Prescriptions]
      summary: Get today's prescriptions
      description: Returns prescribed recovery protocols with rationale
      responses:
        '200':
          description: Today's prescriptions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PrescriptionResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /v1/prescriptions/history:
    get:
      tags: [Prescriptions]
      summary: Get prescription history
      parameters:
        - name: from
          in: query
          schema:
            type: string
            format: date
        - name: to
          in: query
          schema:
            type: string
            format: date
      responses:
        '200':
          description: Prescription history
          content:
            application/json:
              schema:
                type: object
                properties:
                  prescriptions:
                    type: array
                    items:
                      $ref: '#/components/schemas/PrescriptionResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    webhookAuth:
      type: apiKey
      in: header
      name: X-Webhook-Token
  schemas:
...
    SessionSurveyRequest:
      type: object
      required: [sessionAt, stiffness, soreness, mentalReset]
      properties:
        sessionAt:
          type: string
          format: date-time
        stiffness:
          type: integer
          minimum: 1
          maximum: 10
        soreness:
          type: integer
          minimum: 1
          maximum: 10
        mentalReset:
          type: integer
          minimum: 1
          maximum: 10
        notes:
          type: string
          maxLength: 500
        tags:
          type: array
          items:
            type: string
    SessionSurveyResponse:
      type: object
      properties:
        id:
          type: string
        createdAt:
          type: string
          format: date-time
        triggeredRecompute:
          type: boolean
    SessionSurvey:
      allOf:
        - $ref: '#/components/schemas/SessionSurveyRequest'
        - type: object
          properties:
            id:
              type: string
            userId:
              type: string
            createdAt:
              type: string
              format: date-time
    PrescriptionResponse:
      type: object
      properties:
        date:
          type: string
          format: date
        protocols:
          type: array
          items:
            $ref: '#/components/schemas/ProtocolDetail'
        rationale:
          type: object
          properties:
            ruleIds:
              type: array
              items:
                type: string
            keyInputs:
              type: object
              additionalProperties: true
    ProtocolDetail:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        durationMin:
          type: number
        steps:
          type: array
          items:
            type: string
        tags:
          type: array
          items:
            type: string
        equipment:
          type: array
          items:
            type: string
        contraindications:
          type: array
          items:
            type: string
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
              details:
                type: object
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
...